pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dpa-data-cleaning"
dynamic = ["version"]
description = "Data cleaning"
authors = [
  {name = "DP Technology"},
]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3.8",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
]
dependencies = [
    'pydflow>=1.7.14',
    'dpdata',
]
requires-python = ">=3.8"
readme = "README.md"
keywords = ["data cleaning"]

[project.urls]
repository = "https://github.com/smc-lucky/dpa-data-cleaning"

[project.scripts]
dpclean = "dpclean.entrypoints.main:main"

[tool.setuptools.packages.find]
include = ["dpclean*"]

[tool.setuptools_scm]
write_to = "dpclean/_version.py"

[tool.cibuildwheel]
test-command = ["dpclean -h",]
